Feat: game stops after 3 lives and modal works
Now the game stops to run when player loose 3 lives. The start
modal appear now at the beginnig of the game, explaining the goal.
The start button has an event listener that add the allowed keys
to the page by the startGame function. Basically, the keys are not
working before press start, so it's not possible to move the player
by mistake when there's still the modal